==> step1_prepare_news_and_Ads_dataset_1016_2023.ipynb <==
# Data Preparation

Note: This notebook was tested in Amazon SageMaker Studio with Python 3 (Data Science 3.0, Python3) kernel, suggest 2vCPU and 8GiB memory. 

##### Here, we use two datasets from kaggle, it need install kaggle API first, please refer the guideline https://github.com/Kaggle/kaggle-api
##### Download the digital news and content dataset from kaggle by command 'kaggle datasets download -d benjaminawd/new-york-times-articles-comments-2020'
##### Download the ads dataset from kaggle by command 'kaggle datasets download -d sachsene/amazons-advertisements'
##### Save the two zipfiles in your current notebook folder

#### We will prepare two datasets
1. News/articles dataset
2. advertisements dataset


import zipfile

zip_file_path = 'new-york-times-articles-comments-2020.zip'
extracted_folder_path = 'new-york-times-articles-comments-2020'

# Open the zip file
with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
    # Extract all the contents of zip file in current directory
    zip_ref.extractall(extracted_folder_path)

print(f"Successfully extracted the contents to {extracted_folder_path}.")


zip_file_path = 'amazons-advertisements.zip'
extracted_folder_path = 'amazons-advertisements'

# Open the zip file
with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
    # Extract all the contents of zip file in current directory
    zip_ref.extractall(extracted_folder_path)

print(f"Successfully extracted the contents to {extracted_folder_path}.")

# Prepare News/articles dataset

import pandas as pd

# Specify the path to your CSV file
csv_file_path = "new-york-times-articles-comments-2020/nyt-articles-2020.csv"

# Load the CSV file into a DataFrame
df_nyt_articles_2020= pd.read_csv(csv_file_path)




df_nyt_articles_2020.head()

df_nyt_articles_2020.info()

# Check for missing data
missing_data = df_nyt_articles_2020.isnull().sum()

# Display columns with missing data and their respective counts
missing_data.head(20)


df_nyt_articles_2020['subsection'].fillna('Unknown', inplace=True)


# Check for missing data
missing_data = df_nyt_articles_2020.isnull().sum()

# Display columns with missing data and their respective counts
missing_data.head(20)

df_nyt_articles_2020['abstract'].fillna('Unknown', inplace=True)
df_nyt_articles_2020['keywords'].fillna('Unknown', inplace=True)


# Check for missing data
missing_data = df_nyt_articles_2020.isnull().sum()

# Display columns with missing data and their respective counts
missing_data.head(20)

import pandas as pd
import numpy as np

# Add a new column 'np_array' filled with NumPy arrays of 0s
df_nyt_articles_2020['np_array'] = np.zeros(df_nyt_articles_2020.shape[0], dtype=object)

df_nyt_articles_2020.info()

df_nyt_articles_2020.head()

# Save the modified DataFrame to a new CSV file
df_nyt_articles_2020.to_csv("nyt-articles-2020-modified.csv", index=False)


# You can use the head method to get the first 100 records
first_100_records = df_nyt_articles_2020.head(100)

# Save the first 100 records to a JSON file
first_100_records.to_json("nyt-articles-2020-modified-100-records.json", orient="records")


import json

# Convert the DataFrame to a list of dictionaries
data_as_list_of_dicts = df_nyt_articles_2020.to_dict(orient='records')

# Define the output JSON file path
output_json_file = 'nyt_articles_2020.json'

# Write the list of dictionaries to a JSON file
with open(output_json_file, 'w') as json_file:
    json.dump(data_as_list_of_dicts, json_file, indent=4)

print(f'Data has been written to {output_json_file}')


# Preapare Ads dataset


# Specify the path to your CSV file
csv_file_path = "amazons-advertisements/amazon_combined_scrapped_data.csv"

# Load the CSV file into a DataFrame
df_amazon_combined_scrapped_data= pd.read_csv(csv_file_path)

df_amazon_combined_scrapped_data.head()

df_amazon_combined_scrapped_data.info()

# Check for missing data
missing_data_ads = df_amazon_combined_scrapped_data.isnull().sum()

# Display columns with missing data and their respective counts
missing_data_ads.head(10)

# Drop rows with missing data
df_amazon_combined_scrapped_data.dropna(inplace=True)

# Check for missing data
missing_data_ads = df_amazon_combined_scrapped_data.isnull().sum()

# Display columns with missing data and their respective counts
missing_data_ads.head(10)


# Add a new column 'np_array' filled with NumPy arrays of 0s
df_amazon_combined_scrapped_data['np_array'] = np.zeros(df_amazon_combined_scrapped_data.shape[0], dtype=object)

df_amazon_combined_scrapped_data.info()


# Create a new column "ad_id" with unique IDs
df_amazon_combined_scrapped_data['ad_id'] = '2023-' + (df_amazon_combined_scrapped_data.index + 1).astype(str)

# Display the updated DataFrame
print(df_amazon_combined_scrapped_data)


df_amazon_combined_scrapped_data.head()

# Use the sample method to randomly select 5000 records
df_random_sample = df_amazon_combined_scrapped_data.sample(n=5000, random_state=42)

# Save the randomly 5000 records to a JSON file
df_random_sample.to_json("amazon_combined_scrapped_data_5k.json", orient="records")


#### It's a list with 5k ads items, we will use it in notebook practice.

# Save the modified DataFrame to a new CSV file
df_amazon_combined_scrapped_data.to_csv("amazon_combined_scrapped_data_modified.csv", index=False)

#### It's a full list

import json

# Convert the DataFrame to a list of dictionaries
ads_data_as_list_of_dicts = df_amazon_combined_scrapped_data.to_dict(orient='records')

# Define the output JSON file path
ads_output_json_file = 'amazon_combined_scrapped_data.json'

# Write the list of dictionaries to a JSON file
with open(ads_output_json_file, 'w') as json_file:
    json.dump(ads_data_as_list_of_dicts, json_file, indent=4)

print(f'Data has been written to {ads_output_json_file}')
